Create two separate classes for managing Baltech devices.
